Rewrite NotificationHubAdapter to use Volley.
This will reduce the surface area of code we must maintain, while also
giving us immediate support for the Android platform to negotiate TLS
for us.
