--- notification-hubs-sdk/src/main/java/com/microsoft/windowsazure/messaging/notificationhubs/HttpClientNetworkStateHandler.java
@@ -1,104 +1,0 @@
-/*
- * Copyright (c) Microsoft Corporation. All rights reserved.
- * Licensed under the MIT License.
- */
-
-package com.microsoft.windowsazure.messaging.notificationhubs;
-
-import android.util.Log;
-
-import java.io.IOException;
-import java.util.HashSet;
-import java.util.Map;
-import java.util.Set;
-
-/**
- * Decorator pausing calls while network is down.
- */
-class HttpClientNetworkStateHandler extends HttpClientDecorator implements NetworkStateHelper.Listener {
-
-    /**
-     * Network state helper.
-     */
-    private final NetworkStateHelper mNetworkStateHelper;
-
-    /**
-     * All pending calls.
-     */
-    private final Set<Call> mCalls = new HashSet<>();
-
-    /**
-     * Init.
-     *
-     * @param decoratedApi       decorated API.
-     * @param networkStateHelper network state helper.
-     */
-    public HttpClientNetworkStateHandler(HttpClient decoratedApi, NetworkStateHelper networkStateHelper) {
-        super(decoratedApi);
-        mNetworkStateHelper = networkStateHelper;
-        mNetworkStateHelper.addListener(this);
-    }
-
-    @Override
-    public synchronized ServiceCall callAsync(String url, String method, Map<String, String> headers, CallTemplate callTemplate, ServiceCallback serviceCallback) {
-        Call call = new Call(mDecoratedApi, url, method, headers, callTemplate, serviceCallback);
-        if (mNetworkStateHelper.isNetworkConnected()) {
-            call.run();
-        } else {
-            mCalls.add(call);
-            Log.d("ANH", "Call triggered with no network connectivity, waiting network to become available...");
-        }
-        return call;
-    }
-
-    @Override
-    public synchronized void close() throws IOException {
-        mNetworkStateHelper.removeListener(this);
-        mCalls.clear();
-        super.close();
-    }
-
-    @Override
-    public void reopen() {
-        mNetworkStateHelper.addListener(this);
-        super.reopen();
-    }
-
-    @Override
-    public synchronized void onNetworkStateUpdated(boolean connected) {
-        if (connected && mCalls.size() > 0) {
-            Log.d("ANH", "Network is available. " + mCalls.size() + " pending call(s) to submit now.");
-            for (Call call : mCalls) {
-                call.run();
-            }
-            mCalls.clear();
-        }
-
-        /*
-         * Loss of network may not lead to call failure. Also, cancelling the ongoing requests
-         * can cause requests duplication, so let it fail and then retry when we're online again.
-         */
-    }
-
-    private synchronized void cancelCall(Call call) {
-        if (call.mServiceCall != null) {
-            call.mServiceCall.cancel();
-        }
-        mCalls.remove(call);
-    }
-
-    /**
-     * Call wrapper logic.
-     */
-    private class Call extends HttpClientCallDecorator {
-
-        Call(HttpClient decoratedApi, String url, String method, Map<String, String> headers, CallTemplate callTemplate, ServiceCallback serviceCallback) {
-            super(decoratedApi, url, method, headers, callTemplate, serviceCallback);
-        }
-
-        @Override
-        public void cancel() {
-            cancelCall(this);
-        }
-    }
-}
