--- notification-hubs-sdk/src/main/java/com/microsoft/windowsazure/messaging/notificationhubs/ServiceCallback.java
@@ -1,26 +1,0 @@
-/*
- * Copyright (c) Microsoft Corporation. All rights reserved.
- * Licensed under the MIT License.
- */
-
-package com.microsoft.windowsazure.messaging.notificationhubs;
-
-/**
- * The callback used for client side asynchronous operations.
- */
-interface ServiceCallback {
-
-    /**
-     * Implement this method to handle successful REST call results.
-     *
-     * @param httpResponse the HTTP response.
-     */
-    void onCallSucceeded(HttpResponse httpResponse);
-
-    /**
-     * Implement this method to handle REST call failures.
-     *
-     * @param e the exception thrown from the pipeline.
-     */
-    void onCallFailed(Exception e);
-}
